cast wavfile to float32

+++ recipes/qasr/convert_hf_lhotse.py
@@ -29,7 +29,10 @@ def main(args):
         if not os.path.exists(speaker_dir):
             os.makedirs(speaker_dir)
         wav_path = f"{speaker_dir}/{sample['audio']['path']}"
-        write(wav_path, sample['audio']['sampling_rate'], sample['audio']['array'])
+        write(
+            wav_path,
+            sample['audio']['sampling_rate'],
+            sample['audio']['array'].astype(np.float32))
         recording = Recording.from_file(wav_path)
         supervision = SupervisionSegment(
             id=recording.id,
